fix finances

+++ src/app/payment-plan/payment-plan.model.ts
@@ -47,10 +47,10 @@ export class PaymentPlanModel extends Model<PaymentPlanDto> {
   })
   sale_type: string;
 
-  @Column({ type: DataType.DECIMAL(10, 2), allowNull: false })
+  @Column({ type: DataType.DOUBLE(10, 2), allowNull: false })
   total_amount: number;
 
-  @Column({ type: DataType.DECIMAL(10, 2), allowNull: false })
+  @Column({ type: DataType.DOUBLE(10, 2), allowNull: false })
   separation_amount: number;
 
   @Column({ type: DataType.DATEONLY, allowNull: false })
